+++ flutter_local_notifications/lib/src/helpers.dart
@@ -1,7 +1,7 @@
 import 'package:clock/clock.dart';
 import 'package:timezone/timezone.dart';
 
-import 'types.dart';
+import 'package:flutter_local_notifications_platform_interface/flutter_local_notifications_platform_interface.dart';
 
 /// Helper method for validating a date/time value represents a future point in
 /// time where `matchDateTimeComponents` is null.