+++ flutter_local_notifications_windows/lib/src/plugin/stub.dart
@@ -72,8 +72,9 @@ class FlutterLocalNotificationsWindows extends WindowsNotificationsBase {
     String? title,
     String? body,
     required TZDateTime scheduledDate,
-    WindowsNotificationDetails? notificationDetails,
     String? payload,
+    DateTimeComponents? matchDateTimeComponents,
+    WindowsNotificationDetails? notificationDetails,
   }) async {}
 
   @override